+++ ouroboros-consensus-cardano/src/shelley/Ouroboros/Consensus/Shelley/Ledger/NetworkProtocolVersion.hs
@@ -37,6 +37,9 @@ data ShelleyNodeToClientVersion =
 
     -- | New queries introduced: GetConstitutionHash, GetFilteredVoteDelegatees
   | ShelleyNodeToClientVersion8
+
+    -- | New queries introduced: GetProposals, GetRatifyState
+  | ShelleyNodeToClientVersion9
   deriving (Show, Eq, Ord, Enum, Bounded)
 
 instance HasNetworkProtocolVersion (ShelleyBlock proto era) where
@@ -63,6 +66,7 @@ instance SupportedNetworkProtocolVersion (ShelleyBlock proto era) where
       , (NodeToClientV_14, ShelleyNodeToClientVersion6)
       , (NodeToClientV_15, ShelleyNodeToClientVersion7)
       , (NodeToClientV_16, ShelleyNodeToClientVersion8)
+      , (NodeToClientV_17, ShelleyNodeToClientVersion9)
       ]
 
   latestReleasedNodeVersion = latestReleasedNodeVersionDefault